+++ src/headless/plugins/bookmarks/model.js
@@ -8,6 +8,13 @@ class Bookmark extends Model {
         return 'jid';
     }
 
+    /**
+     * @returns {boolean}
+     */
+    get pinned() {
+        return this.get('extensions')?.some(/** @param {String} e */ e => e.includes('<pinned') && e.includes('urn:xmpp:bookmarks-pinning:0'));
+    }
+
     getDisplayName() {
         return this.get('name') && Strophe.xmlunescape(this.get('name')) || this.get('jid');
     }