+++ CHANGES.md
@@ -15,6 +15,7 @@
 - #3939: Don't show invitations to groupchats in which the user is already present
 - Don't render unfurls for retracted messages.
 - #3941: add adhoc completed command result and text-multi as merged lines of text
+- #3949: Allow pinning bookmarked conversations to the top (XEP-0469)
 
 ## 12.0.0 (2025-08-28)